backend/requirements.txt
flask>=2.0.0
flask-socketio>=5.0.0
flask-cors>=3.0.0
python-socketio>=5.0.0
eventlet>=0.33.0
pymysql>=1.0.0
requests>=2.28.0
PyJWT>=2.8.0
zhipuai>=2.0.0
